requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
beautifulsoup4==4.14.2
bitsandbytes==0.47.0
blake3==1.0.8
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
braceexpand==0.1.7
Brotli==1.1.0
cachetools==6.2.1
catalogue==2.0.10
cbor2==5.7.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.2.1
cloudpathlib==0.23.0
cloudpickle==3.1.1
compressed-tensors==0.11.0
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.2
cupy-cuda12x==13.6.0
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
decord==0.6.0
deepspeed==0.14.4
depyf==0.19.0
diffusers==0.29.2
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fairscale==0.4.4
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
ffmpy==0.6.1
filelock==3.20.0
flash_attn==2.8.3
fonttools==4.60.0
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
ftfy==6.3.1
gdown==5.2.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
gradio==5.49.1
gradio_client==1.13.3
graphviz==0.21
groovy==0.1.2
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.24.7
identify==2.6.15
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
inquirerpy==0.3.4
interegular==0.3.3
iopath==0.1.10
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
lazy_loader==0.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
murmurhash==1.0.13
narwhals==2.8.0
networkx==3.4.2
ninja==1.13.0
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==2.3.0
openai-harmony==0.0.4
opencv-python==4.12.0.88
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.11.3
outlines_core==0.2.11
packaging==25.0
pandas==2.3.2
parso==0.8.5
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pexpect==4.9.0
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
portalocker==3.2.0
pre_commit==4.3.0
preshed==3.0.10
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pycocoevalcap==1.2
pycocotools==2.0.10
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.6
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-magic==0.4.27
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.50.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.1
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.38.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.11.0
timm==0.4.12
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
torchviz==0.0.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.45.2
transformers-stream-generator==0.0.5
triton==3.4.0
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
uvloop==0.21.0
virtualenv==20.35.3
wandb==0.22.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.14
weasel==0.4.1
webdataset==1.0.2
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xformers==0.0.32.post1
xgrammar==0.1.25
yarl==1.22.0
zipp==3.23.0
